Broke configuration manager into smaller pieces, to enhance maintainability
